+++ CHANGELOG.md
@@ -4,8 +4,8 @@ This project follows [Semantic Versioning](http://semver.org/).
 
 ## [41.3.0] 2026-04-02
 ### Added
-- New utility function validateQuestionnaireOnFHIRServer() to validate a
-  questionnaire against a FHIR server.
+- New utility function LForms.Util.validateQuestionnaireOnFHIRServer()
+  to validate a questionnaire against a FHIR server.
 
 ## [41.2.0] 2026-04-02
 ### Added